Fixes gclone function for zsh
The format for read is different between bash and zsh. With the -p flag,
zsh expets to receive a number of processes to use. To fix this, I have
them each read the important bits in their own way before passing that
on to the main function. In order to get around the read at the end, I
used fzf which I think I like better anyway.